# sdd/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 任务管理系统 API
  description: |
    任务管理系统后端 API 文档
    
    核心功能：
    - Excel 导入任务数据
    - 模板管理
    - Rundown 发布清单管理
    - 任务执行控制
    - Jenkins/Ansible 自动化集成
    - 定时任务调度
    - 批量执行
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发服务器
  - url: http://{host}:{port}/api/v1
    description: 生产服务器
    variables:
      host:
        default: localhost
      port:
        default: "8000"
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: 用户登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: 用户名
                password:
                  type: string
                  description: 密码
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: 用户登出
      operationId: logout
      security:
        - CookieAuth: []
      responses:
        '200':
          description: 登出成功

  # ==================== Import ====================
  /import/upload:
    post:
      tags:
        - Import
      summary: 上传并解析 Excel 文件
      operationId: uploadExcel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Excel 文件 (.xlsx, .xls)
      responses:
        '200':
          description: 解析成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '400':
          description: 文件格式错误或解析失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /import/validate:
    post:
      tags:
        - Import
      summary: 验证解析后的数据
      operationId: validateImportData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - data
              properties:
                data:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: 验证结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'

  # ==================== Templates ====================
...
  /rundowns/{rundown_id}/execute:
    post:
      tags:
        - Execution
      summary: 执行 Rundown
      operationId: executeRundown
      parameters:
        - name: rundown_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: 开始执行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
  /rundowns/batch-execute:
    post:
      tags:
        - Execution
      summary: 批量执行 Rundown
      operationId: batchExecuteRundowns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rundown_ids
              properties:
                rundown_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '202':
          description: 开始批量执行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchExecutionResponse'

  # ==================== Tasks ====================
...
  /tasks/{task_id}/execute:
    post:
      tags:
        - Execution
      summary: 执行单个任务
      operationId: executeTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: 开始执行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
...
  /config:
    get:
      tags:
        - Config
      summary: 获取系统配置列表
      operationId: listConfig
      responses:
        '200':
          description: 配置列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SystemConfig'

    post:
      tags:
        - Config
      summary: 创建/更新系统配置
      operationId: setConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetConfigRequest'
      responses:
        '200':
          description: 设置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfig'

# ==================== Components ====================
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
...
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        role:
          type: string

    # Import
    ImportResult:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
        total_rows:
          type: integer
    ValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              field:
                type: string
              message:
                type: string

    # Templates
...
    ExecutionResponse:
      type: object
      properties:
        execution_id:
          type: string
        message:
          type: string
    BatchExecutionResponse:
      type: object
      properties:
        batch_id:
          type: string
        total:
          type: integer
        message:
          type: string

    # Scheduling
...
    SystemConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        config_key:
          type: string
        value_type:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SetConfigRequest:
      type: object
      required:
        - config_key
        - config_value
      properties:
        config_key:
          type: string
        config_value:
          type: string
        value_type:
          type: string
          default: string
        description:
          type: string

    # Error
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string